requirements.txt
docopt
flask
flask_uploads
python-docx
